add authentication

+++ resources/collection_requests.py
@@ -5,6 +5,7 @@ Blueprint for handling collection requests
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import jwt_required, get_jwt
 
 from db import db
 from models import CollectionRequestModel
@@ -23,6 +24,7 @@ class CollectionRequests(MethodView):
     """
     Class for handling requests to the /collection_requests endpoint
     """
+    @jwt_required()
     @blp.response(200, CollectionRequestSchema(many=True))
     def get(self):
         """
@@ -32,8 +34,12 @@ class CollectionRequests(MethodView):
             dict: A dictionary containing all collection requests in
             the database
         """
-        return CollectionRequestModel.query.all()
+        jwt = get_jwt()
+        if jwt.get("role") in ("admin", "household"):
+            return CollectionRequestModel.query.all()
+        abort(403, message="Admin or household privileges required to access resources")
 
+    @jwt_required()
     @blp.arguments(CollectionRequestSchema)
     @blp.response(201, CollectionRequestSchema)
     def post(self, collection_request_data):
@@ -51,8 +57,11 @@ class CollectionRequests(MethodView):
             abort(400, message): If there is an error adding the collection
             request to the database
         """
-        collection_request = CollectionRequestModel(**collection_request_data)
+        jwt = get_jwt()
+        if jwt.get("role") != "household":
+            abort(403, message="Household privileges required to access resources")
 
+        collection_request = CollectionRequestModel(**collection_request_data)
         try:
             db.session.add(collection_request)
             db.session.commit()
@@ -69,6 +78,7 @@ class CollectionRequest(MethodView):
     Class for handling requests to the
     /collection_requests/<collection_request_id> endpoint
     """
+    @jwt_required()
     @blp.response(200, CollectionRequestSchema)
     def get(self, collection_request_id):
         """
@@ -85,8 +95,12 @@ class CollectionRequest(MethodView):
             NotFound: If the collection request with the given ID does
             not exist
         """
-        return CollectionRequestModel.query.get_or_404(collection_request_id)
+        jwt = get_jwt()
+        if jwt.get("role") in ("admin", "household"):
+            return CollectionRequestModel.query.get_or_404(collection_request_id)
+        abort(403, message="Admin or household privileges required to access resources")
 
+    @jwt_required()
     def delete(self, collection_request_id):
         """
         Delete a collection request by ID
@@ -103,9 +117,12 @@ class CollectionRequest(MethodView):
             NotFound: If the collection request with the given ID does
             not exist
         """
+        jwt = get_jwt()
+        if jwt.get("role") != "admin":
+            abort(403, message="Admin privileges required to delete a collection request")
+
         collection_request = CollectionRequestModel.query.get_or_404(
             collection_request_id)
         db.session.delete(collection_request)
         db.session.commit()
-
         return {"message": "Collection request deleted successfully."}
